Catch error and rename package to not collide with others
