+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,3 @@
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactlibrary">
-
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.rninstallreferrer">
 </manifest>
