+++ android/src/main/java/com/rninstallreferrer/RNInstallReferrerPackage.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.rninstallreferrer;
 
 import java.util.Arrays;
 import java.util.Collections;
