+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-install-referrer",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "description": "",
   "homepage": "https://github.com/Gretchin/react-native-install-referrer#readme",
   "license": "MIT",
